final-pjt-back/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2
django-allauth==0.50.0
django-appconf==1.0.5
django-cors-headers==3.13.0
django-extensions==3.2.1
django-imagekit==4.1.0
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.0
lxml==4.9.1
oauthlib==3.2.2
pilkit==2.0
Pillow==9.3.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
